requirements.txt
pandas~=2.2.3
aiogram~=3.0.0b5
python-dotenv~=1.0.1
aiohttp~=3.11.14
lxml~=5.3.1
fake-useragent~=1.1.3
selenium~=4.8.0